refactor: bzCurve 분리
